+++ v5/io_dest_test.go
@@ -280,7 +280,7 @@ func TestDestImplementsReadBuffer(t *testing.T) {
 	// ----------------------------------------------------------------
 	// perform the change
 
-	_, ok := i.(ReadBuffer)
+	_, ok := i.(Input)
 
 	// ----------------------------------------------------------------
 	// test the results